Add a config setting to disable use of GPIO pins.
This is useful if the Pi does not have the custom board installed,
and you want to prevent it from interacting with pins that are
in use for other purposes.

Split tests into hermetic unit tests which do not interact with
hardware, and integration tests which exercise real modem and
indicator hardware.